Add fluentd logging - and change k8s.labels.app to *.application

+++ deploy/src/clean.sh
@@ -31,5 +31,12 @@ if [ -d "../services/" ]; then
 	rm -rf ../services/;
 fi
 
-rm  01_app-config-blue.yaml
-rm  01_app-config-green.yaml
+if test -f "01_app-config-blue.yaml"; then
+	echo "deleteing 01_app-config-blue.yaml"
+	rm  01_app-config-blue.yaml
+fi
+
+if test -f "01_app-config-green.yaml"; then
+	echo "deleteing 01_app-config-green.yaml"
+	rm  01_app-config-green.yaml
+fi
